backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-rag-backend"
version = "0.1.0"
description = "FastAPI backend for PDF RAG chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "pypdf>=4.3.0",
  "sentence-transformers>=3.0.1",
  "faiss-cpu>=1.8.0.post1",
  "groq>=0.11.0",
  "numpy>=1.26.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.2",
  "ruff>=0.6.8"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
